+++ Editor/BitmapFontCreator.cs
@@ -80,29 +80,32 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 			var font = new Font(baseName)
 			{
 				material = material,
-				characterInfo = CreateCharacters(data, map),
+				characterInfo = CreateCharacters(data, map, data.Descent),
 			};
 
 			var so = new SerializedObject(font);
 			so.FindProperty("m_LineSpacing").floatValue = data.LineSpacing;
+			so.FindProperty("m_Ascent").floatValue = data.Ascent + data.Descent;
+			so.FindProperty("m_Descent").floatValue = data.Descent;
 			so.ApplyModifiedProperties();
 
 			return font;
 		}
 
-		private static CharacterInfo[] CreateCharacters(ExecutionData data, Dictionary<char, CharacterProps> map)
+		private static CharacterInfo[] CreateCharacters(ExecutionData data, Dictionary<char, CharacterProps> map, float descent)
 		{
 			var texSize = new Vector2Int(data.Texture.width, data.Texture.height);
 			var cellSize = new Vector2Int(Mathf.FloorToInt(texSize.x / data.Cols), Mathf.FloorToInt(texSize.y / data.Rows));
 			var cellUVSize = new Vector2(1f / data.Cols, 1f / data.Rows);
+			var ratio = new Vector2(1f / texSize.x, 1f / texSize.y);
 
 			var characters = new List<CharacterInfo>();
-			int xMin, xMax, advance, w, advMax = 0;
+			int xMin, xMax, yMin, yMax, advance, w, advMax = 0;
 
 #if BITMAP_FONT_CREATOR_DEBUG
 			static string _(float x) => $"<color=yellow>{x}</color>";
 #endif
-
+			var baseline = (int)(cellSize.y - descent);
 			for (var row = 0; row < data.Rows; row++)
 			{
 				for (var col = 0; col < data.Cols; col++)
@@ -121,7 +124,9 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 						width: cellSize.x,
 						height: cellSize.y,
 						xMin: out xMin,
-						xMax: out xMax
+						xMax: out xMax,
+						yMin: out yMin,
+						yMax: out yMax
 					);
 
 					w = xMax - xMin;
@@ -129,16 +134,26 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 					advance = w + data.DefaultCharacterSpacing;
 					if (advance > advMax) advMax = advance;
 
+					// var y = Mathf.RoundToInt(-ascent + descent + (yMax - yMin));
+					var y = Mathf.RoundToInt(-yMin + descent - (yMax - baseline));
+
+
 					var info = new CharacterInfo
 					{
 						index = ch,
-						uvTopLeft = new Vector2(cellUVSize.x * col, cellUVSize.y * (data.Rows - row - 1)),
-						uvBottomRight = new Vector2(cellUVSize.x * (col + 1), cellUVSize.y * (data.Rows - row)),
-						minX = 0,
-						maxX = Mathf.RoundToInt(cellSize.x),
-						minY = Mathf.RoundToInt(cellSize.y * 0.5f),
-						maxY = Mathf.RoundToInt(-cellSize.y * 0.5f),
-						bearing = -xMin,
+						uvTopLeft = new Vector2(
+							cellUVSize.x * col + (xMin * ratio.x),
+							cellUVSize.y * (data.Rows - row) - (yMin * ratio.y)
+						),
+						uvBottomRight = new Vector2(
+							cellUVSize.x * (col + 1) - ((cellSize.x - xMax) * ratio.x),
+							cellUVSize.y * (data.Rows - row - 1) + ((cellSize.y - yMax) * ratio.y)
+						),
+						minX = xMin,
+						maxX = xMax,
+						minY = yMin + y,//0
+						maxY = yMax + y,//cellSize.y,
+						bearing = 0,
 						advance = advance,
 					};
 
@@ -154,7 +169,7 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 					characters.Add(info);
 
 #if BITMAP_FONT_CREATOR_DEBUG
-					Debug.Log($"<b>{ch}</b> xMin: {_(xMin)} xMax: {_(xMax)} advance: {_(advance)}");
+					Debug.Log($"<b>{ch}</b> {_(info.glyphWidth)} {_(info.glyphHeight)} yMin: {_(yMin)} yMax: {_(yMax)}");
 #endif
 				}
 			}
@@ -166,12 +181,12 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 		}
 
 		private static void GetCharacterBounds(Texture2D tex, float alphaThreshold, int x0, int y0,
-			int width, int height, out int xMin, out int xMax)
+			int width, int height, out int xMin, out int xMax, out int yMin, out int yMax)
 		{
 			xMin = width;
 			xMax = 0;
-			// yMin = height;
-			// yMax = 0;
+			yMin = height;
+			yMax = 0;
 
 			int xx, yy;
 			for (var y = 0; y < height; y++)
@@ -183,8 +198,8 @@ namespace dev.klebersilva.tools.bitmapfontcreator
 					if (tex.GetPixel(xx, yy).a <= alphaThreshold) continue;
 					if (x < xMin) xMin = x;
 					if (x > xMax) xMax = x;
-					// if (y < yMin) yMin = y;
-					// if (y > yMax) yMax = y;
+					if (y < yMin) yMin = y;
+					if (y > yMax) yMax = y;
 				}
 			}
 		}